requirements.txt
streamlit==1.8.1
opencv-python==4.5.5.64
Pillow==9.0.1
numpy==1.22.2
pandas==1.4.1
